requirements.txt
# Core Dependencies
falkordb>=1.2.0
falkordb-orm>=1.2.0
graphiti-core>=0.3.0
python-dotenv>=1.0.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# LLM & RAG
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.3.0
tiktoken>=0.5.0

# Graph Analytics
# FalkorDB provides native graph algorithms (PageRank, Betweenness, Label Propagation, BFS)
# No external graph libraries needed

# Spatial & Geospatial
shapely>=2.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Database & Caching
# Redis and SQLAlchemy not currently used

# API
# GraphQL and httpx not currently used

# Utilities
pyyaml>=6.0.0
python-dateutil>=2.8.0
pytz>=2023.3

# Logging & Monitoring
# Using stdlib logging; Prometheus not currently used

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
hypothesis>=6.92.0

# Development
black>=23.11.0
ruff>=0.1.6
mypy>=1.7.0
pre-commit>=3.5.0

# Documentation
# MkDocs not currently used
